backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.15.2
pydantic-settings==2.9.1
email-validator==2.2.0
mcp==1.9.2
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
cryptography==44.0.3
httpx==0.28.1
sse-starlette==2.2.1
apscheduler==3.11.0
python-multipart==0.0.20
